configure_esp_secure_cert.py: fix efuse_key_file argument configuration
The call to `add_argument` that adds this option has `nargs=1` as an
input parameter, which means that the field `efuse_key_file` of the
`Namespace` object returned by `parse_args` will be a list.

This results in an error when `efuse_key_file` is passed to
`os.path.exists()`, as this function does not expect a list.

Removing the `nargs=1` parameter from `add_argument` should fix this
issue.